move load-data worker into a new folder structure
